Tweak swift loading view to show AliasVault logo (#1562)
