+++ apps/mobile-app/ios/Autofill/CredentialProviderViewController.swift
@@ -70,10 +70,9 @@ public class CredentialProviderViewController: ASCredentialProviderViewControlle
         super.viewWillAppear(animated)
 
         // Check if we're in quick return mode
-        // Don't show any UI - Face ID will appear immediately with no background
+        // Show branded loading view while Face ID authentication is in progress
         if isQuickReturnMode {
-            // Make the view transparent so system UI shows through
-            view.backgroundColor = .clear
+            showBrandedLoadingView()
             return
         }
 
@@ -202,6 +201,25 @@ public class CredentialProviderViewController: ASCredentialProviderViewControlle
                 ))
             }
         }
+    }
+
+    /// Show branded loading view with AliasVault logo during biometric authentication
+    private func showBrandedLoadingView() {
+        let loadingView = BrandedLoadingView(message: nil, showLoadingAnimation: false)
+        let hostingController = UIHostingController(rootView: loadingView)
+        addChild(hostingController)
+        view.addSubview(hostingController.view)
+
+        hostingController.view.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            hostingController.view.topAnchor.constraint(equalTo: view.topAnchor),
+            hostingController.view.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            hostingController.view.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            hostingController.view.bottomAnchor.constraint(equalTo: view.bottomAnchor)
+        ])
+
+        hostingController.didMove(toParent: self)
+        currentHostingController = hostingController
     }
 
     /// Show PIN unlock for quick return mode when biometric fails
