+++ apps/mobile-app/ios/VaultUI/Bundle+VaultUI.swift
@@ -2,8 +2,8 @@ import Foundation
 
 private final class VaultUIBundleToken {}
 
-extension Bundle {
-    /// The bundle that contains VaultUI’s localized resources
+public extension Bundle {
+    /// The bundle that contains VaultUI's localized resources
     static var vaultUI: Bundle {
         // If you don’t use a separate .bundle target, this is the framework bundle itself.
         let frameworkBundle = Bundle(for: VaultUIBundleToken.self)
