+++ apps/mobile-app/ios/VaultUI/Registration/Passkeys/PasskeyRegistrationView.swift
@@ -42,7 +42,7 @@ public struct PasskeyRegistrationView: View {
 
                 // Loading overlay
                 if viewModel.isLoading {
-                    LoadingOverlayView(message: viewModel.loadingMessage)
+                    BrandedLoadingView(message: viewModel.loadingMessage)
                 }
             }
             .navigationBarTitleDisplayMode(.inline)
